functions/predict_race_get/requirements.txt
# Function dependencies, for example:
# package>=version
beautifulsoup4==4.8.2
requests==2.23.0
google-cloud-bigquery==1.24.0
google-cloud-pubsub==1.4.2
google-cloud-storage==1.26.0
